update with moz 1.9.2/FF 3.6 release

+++ jsscan.cpp
@@ -835,7 +835,7 @@ NewToken(JSTokenStream *ts, ptrdiff_t adjust)
     tp->pos.begin.index = ts->linepos +
                           (tp->ptr - ts->linebuf.base) -
                           ts->ungetpos;
-    tp->pos.begin.lineno = tp->pos.end.lineno = (uint16)ts->lineno;
+    tp->pos.begin.lineno = tp->pos.end.lineno = ts->lineno;
     return tp;
 }
 
@@ -914,7 +914,7 @@ js_GetToken(JSContext *cx, JSTokenStream *ts)
             if (!atom)
                 goto error;
         }
-        tp->pos.end.lineno = (uint16)ts->lineno;
+        tp->pos.end.lineno = ts->lineno;
         tp->t_op = JSOP_STRING;
         tp->t_atom = atom;
         goto out;
@@ -1019,7 +1019,7 @@ js_GetToken(JSContext *cx, JSTokenStream *ts)
             atom = atomize(cx, tb);
             if (!atom)
                 goto error;
-            tp->pos.end.lineno = (uint16)ts->lineno;
+            tp->pos.end.lineno = ts->lineno;
             tp->t_op = JSOP_STRING;
             tp->t_atom = atom;
             tt = TOK_XMLATTR;
@@ -1289,7 +1289,7 @@ retry:
         atom = atomize(cx, tb);
         if (!atom)
             goto error;
-        tp->pos.end.lineno = (uint16)ts->lineno;
+        tp->pos.end.lineno = ts->lineno;
         tp->t_op = JSOP_STRING;
         tp->t_atom = atom;
         tt = TOK_STRING;
@@ -1523,7 +1523,7 @@ retry:
                 if (!atom)
                     goto error;
                 tp->t_atom = atom;
-                tp->pos.end.lineno = (uint16)ts->lineno;
+                tp->pos.end.lineno = ts->lineno;
                 goto out;
             }
